Make i2c function signitures consistent
The names of the i2c functions in ioboard weren't consistent with the
naming scheme, and the types weren't C99 fixed width. There was also an
issue with ints being passed as buffer pointers. I also added some
functions to read and write single bytes, to avoid having to set up
buffers for small operations, and functions for reading from a register
by writing the address of the register first.