Most of the mbed lib was unused, simplify to lib called time
The mbed library was intended to have a bunch of functions in it that only
needed the mbed to function (i.e, they didn't have the generality of CMSIS by
assuming having the mbed). However, none were used in this project except for
the time related functionality. I've decided to break this out into it's own,
simpler library, just called time.

This commit also makes sure all the existing files use the new names.